requirements.txt
absl-py==0.1.9
astor==0.6.2
backports.functools-lru-cache==1.4
backports.weakref==1.0.post1
bleach==1.5.0
cycler==0.10.0
decorator==4.2.1
enum34==1.1.6
funcsigs==1.0.2
futures==3.1.1
gast==0.2.0
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
Keras==2.1.5
Markdown==2.6.11
matplotlib==2.1.2
mock==2.0.0
networkx==2.1
numpy==1.14.0
opencv-python==3.4.0.12
pandas==0.22.0
pbr==3.1.1
Pillow==5.0.0
protobuf==3.5.1
pyparsing==2.2.0
python-dateutil==2.6.1
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
scikit-image==0.13.1
scikit-learn==0.19.1
scipy==1.0.0
six==1.11.0
tensorboard==1.6.0
tensorflow-gpu==1.6.0
tensorflow-tensorboard==1.5.0
termcolor==1.1.0
tqdm==4.19.9
Werkzeug==0.14.1
